requirements.txt
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
PyMySQL
autopep8
pycodestyle
pandas
numpy
sklearn
gensim
nltk
scipy
